Add sshd feature to dev container
This commit adds support for the sshd feature to the dev
container. Enabling this allows the dev container to be accessed via
SSH directly (useful for non-VS Code environments).

Additionally, this commit fixes an issue in retrieving the Helm
signing key that was causing the signing key to be printed directly to
the terminal, causing various problems.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>